Switch to original ISC license
The forked license appears to be a custom version of ISC with a
simplified warranty disclaimer. By reading through the original
and this version of the wording, one can reasonably conclude that
the original disclaimer covers everything that the simplified
version does and thus switching to the original should not be a
problem.

Speaking of licenses, there are two source files that are covered
by other licenses:

bsdstubs.c   -- OpenBSD ISC (without "and/or")
win-dirent.h -- MIT